# server/api/openapi.yaml
openapi: "3.0.2"
info:
  title: "ERP.AERO"
  description: "Test task"
  version: "1.0.0"
servers:
  - url: "http://localhost:3000/"
security: 
 - Bearer: []        
paths:
...
  /file/list:
    get:
      tags:
      - storage
      description: "get the file list"
      operationId: get the file list
      parameters:
      - name: list_size
        in: query
        description: "number of records on the page"
        schema:
          type: number
          default: 10
          
      - name: page
        in: query
        description: "number of the list page"
        schema:
          type: number
          default: 1
        example: 1
        required: false
        
      responses:
        200:
          description: the file list was returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fileList'

        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
...
components:
  schemas:
...
    file:
      type: object
      properties:
        title:
          type: string
        extension:
          type: string
        mimeType:
          type: string
        size:
          type: string                  
      example: {title: "picture", extension: "png", mimeType: "image", size: "1.2 mb"}   
    fileList:
      type: array
      description: card array
      items:
        $ref: '#/components/schemas/file'
...
  responses:
    400:
      description: "#400. bad request"
      content:
        text/plain:
          schema:
            title: "Bed request"
            type: string
            example: Bed request
    404:
      description: "#404. the endpoint was not found"
      content:
        text/plain:
          schema:
            title: endpoint not found
            type: string
            example: Not found
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: Requests should pass an api_key header.
...